Refactor discard_iterator
